requirements.txt
langchain
openai
langchain_openai
huggingface_hub
python-dotenv
streamlit